+++ tests/Nelmio/Alice/Mocks/FooProvider.php
@@ -9,12 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Nelmio\Alice\Provider;
+namespace Nelmio\Alice\Mocks;
 
-class IdentityProvider
+class FooProvider
 {
-    public static function identity($val)
+    public static function foo($str)
     {
-        return $val;
+        return 'foo' . $str;
     }
 }